2.0.1 - Definities voor CPT verplaatst naar plugin.

+++ front-page.php
@@ -9,8 +9,8 @@
 // * @author  Paul van Buuren
 // * @license GPL-2.0+
 // * @package wp-rijkshuisstijl
-// * @version 1.2.3
-// * @desc.   Responsive styling homepage.
+// * @version 2.0.1
+// * @desc.   Definities voor CPT verplaatst naar plugin.
 // * @link    https://github.com/ICTU/digitale-overheid-wordpress-theme-rijkshuisstijl
 // 
  */
@@ -67,49 +67,59 @@ function rhswp_home_onderwerpen_dossiers() {
 
   }
 
+
+
   echo '<section class="onderwerpen">';
   echo '<div class="wrap">';
-    
-  if( have_rows( 'home_onderwerpen_dossiers' ) ) {
 
-    echo '<div class="row ' . $breedte . '">';
-
-    while( have_rows( 'home_onderwerpen_dossiers') ): the_row(); 
-    
-      $rowcounter++;
-
-      $url_extern   = get_sub_field('kies_een_onderwerp');
-      $acfid        = RHSWP_CT_DOSSIER . '_' . $url_extern->term_id;
-      $kortebeschr  = get_field( 'dossier_korte_beschrijving_voor_dossieroverzicht', $acfid );
-      $description  = $url_extern->description;
-
-      if ( 'standaardbeschrijving' != get_sub_field( 'welke_beschrijving' ) ) {
-        $description = get_sub_field( 'andere_beschrijving' );
-      }
-      elseif ( $kortebeschr ) {
-        $description = $kortebeschr;
-      }
-
-// to do
-
-//WOEHOE!! Waar isme dinges?      
-      
-
-      $name = 'naam';
-      $url  = '';
-      if ( $url_extern ) {
-        $name = $url_extern->name;
-        $url  = get_term_link( $url_extern->term_id );
-      }
-    
-      echo '<a href="' . $url . '" class="linkblock"><h3>' .  $name . '</h3><p>' .  wp_strip_all_tags( $description ) . '</p></a>';
-
-    endwhile; 
-
-    echo '</div>';
-
-
+  if ( ! taxonomy_exists( RHSWP_CT_DOSSIER ) ) {
+    echo __( 'Taxonomie voor dossier bestaat niet. Zet de plugin hiervoor aan ("ICTU / WP Register post types and taxonomies").', 'wp-rijkshuisstijl' );
   }
+
+  if ( taxonomy_exists( RHSWP_CT_DOSSIER ) ) {
+      
+    if( have_rows( 'home_onderwerpen_dossiers' ) ) {
+  
+      echo '<div class="row ' . $breedte . '">';
+  
+      while( have_rows( 'home_onderwerpen_dossiers') ): the_row(); 
+      
+        $rowcounter++;
+  
+        $url_extern   = get_sub_field('kies_een_onderwerp');
+        $acfid        = RHSWP_CT_DOSSIER . '_' . $url_extern->term_id;
+        $kortebeschr  = get_field( 'dossier_korte_beschrijving_voor_dossieroverzicht', $acfid );
+        $description  = $url_extern->description;
+  
+        if ( 'standaardbeschrijving' != get_sub_field( 'welke_beschrijving' ) ) {
+          $description = get_sub_field( 'andere_beschrijving' );
+        }
+        elseif ( $kortebeschr ) {
+          $description = $kortebeschr;
+        }
+  
+  // to do
+  
+  //WOEHOE!! Waar isme dinges?      
+        
+  
+        $name = 'naam';
+        $url  = '';
+        if ( $url_extern ) {
+          $name = $url_extern->name;
+          $url  = get_term_link( $url_extern->term_id );
+        }
+      
+        echo '<a href="' . $url . '" class="linkblock"><h3>' .  $name . '</h3><p>' .  wp_strip_all_tags( $description ) . '</p></a>';
+  
+      endwhile; 
+  
+      echo '</div>';
+  
+  
+    }
+  }
+  
   if ( is_active_sidebar( RHSWP_HOME_WIDGET_AREA ) ) {
 
     dynamic_sidebar( RHSWP_HOME_WIDGET_AREA );
